Fixed files not included in release. Moved node relevant stuff in libloki into its own modules folder.
