+++ libloki/modules/sc_reduce32.js
