+++ package.json
@@ -231,8 +231,7 @@
       "permissions_popup.html",
       "debug_log.html",
       "_locales/**",
-      "libloki/mnemonic.js",
-      "libloki/sc_reduce32.js",
+      "libloki/modules/*.js",
       "mnemonic_languages/**",
       "protos/*",
       "js/**",
